Update hiring-process.md
Preparing to add more details on the operational aspects of hiring, I came across 'HRM' which, even though that falls under the Operations Coordinators tasks, it's not a pure HRM role and thus I changed it to 'OPS'. Boris once told me he doesn't want to be called 'CEO' but 'CE', simply Chief Executive, so I changed that as well. I also added a link to our LinkedIn page.